Linting fixes and more debugging
